+++ packages/unplugin-uni-app-tailwind/tsup.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'tsup';
 
 export default defineConfig([
   {
-    entry: ['src/index.ts'],
+    entry: ['src/*.ts'],
     format: 'esm',
     dts: true,
     minify: true,
@@ -15,7 +15,7 @@ export default defineConfig([
     },
   },
   {
-    entry: ['src/index.ts'],
+    entry: ['src/*.ts'],
     format: 'cjs',
     minify: true,
     shims: true,